Mortgage_Rag/requirements.txt
pypdf==5.1.0
openai==1.55.3
numpy<2.0,>=1.23
pydantic==2.10.6
python-dotenv==1.0.1
streamlit==1.41.1
langchain==0.3.14
langgraph==0.5.3
langchain-openai==0.2.11
langchain-community==0.3.14
langchain-text-splitters>=0.3.3
faiss-cpu==1.13.2
reportlab==4.2.5
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
